refactor: 패키지 이름 변경
- oauth 사용자 정보가 담긴 패키지들로 이루어져 있어 userinfo로 변경
