+++ src/main/java/com/petitapetit/miml/domain/auth/oauth/OAuth2Provider.java
@@ -1,4 +1,4 @@
-package com.petitapetit.miml.domain.auth.oauth.provider;
+package com.petitapetit.miml.domain.auth.oauth;
 
 import java.util.Arrays;
