+++ src/main/java/com/petitapetit/miml/domain/auth/oauth/userinfo/OAuth2UserInfo.java
@@ -1,8 +1,10 @@
-package com.petitapetit.miml.domain.auth.oauth.provider;
+package com.petitapetit.miml.domain.auth.oauth.userinfo;
 
 // OAuth2.0 제공자들 마다 응답해주는 속성값이 달라서 공통으로 만들어준다.
 
 import java.util.Map;
+
+import com.petitapetit.miml.domain.auth.oauth.OAuth2Provider;
 
 public abstract class OAuth2UserInfo {
