+++ src/main/java/com/petitapetit/miml/domain/member/repository/MemberRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.petitapetit.miml.domain.auth.oauth.provider.OAuth2Provider;
+import com.petitapetit.miml.domain.auth.oauth.OAuth2Provider;
 import com.petitapetit.miml.domain.member.model.Member;
 
 @Repository
